first exercise done :)

+++ diff.c
@@ -1,17 +1,18 @@
 //---------------------------------------------------------------
-// Program lab_6a - Asemblery Laboratorium IS II rok
+// Kolos 2015/16 CZW 11:00
 //
-// To compile&link: gcc -o lab_6a lab_6a.c lab_6a_asm.s
-// To run:          ./lab_6a
+// To compile&link: gcc -o diff diff.c diff.s
+// To run:          ./diff
 //
 //---------------------------------------------------------------
 
 #include <stdio.h>
 
-long max_diff(int a, int b, int c, int d);
+long max_diff(long a, long b, long c, long d);
 
 int main( void )
 {
+	//Running few tests
 	printf("max_diff(111,222,333,444) = %ld\n",max_diff(111,222,333,444));
 	printf("max_diff(-1,2,3,4) = %ld\n",max_diff(-1,2,3,4));
 	printf("max_diff(1,2,3,4) = %ld\n",max_diff(1,2,3,4));
